requirements.txt
requests==2.28.1
Markdown==3.3.7
Scrapy==2.6.1
html5lib==1.1
Twisted==22.4.0
feedparser
func_timeout==4.3.5
jsonpath==0.82
django==2.2.28
django-simpleui==2022.6.131
django-apscheduler==0.3.0
django-import-export==2.7.1
playwright